requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydantic[email]
python-jose

elasticsearch

scikit-learn
numpy
pandas

langgraph
langchain
langchain-core

groq

httpx
requests

python-dotenv
python-multipart

# Authentication
python-jose[cryptography]
werkzeug

pytest
black
isort
